Adding the main files, created a bot restart command.
Created a bot restart command that can only be ran by the user id that is defined in the owner_id part of the config. Coded basic bot startup code along with running through the cogs folder on startup. Refer to config_example.py to find out how to setup.