Update moderation.py
Added 2 new commands, kick which just kicks the specified member from the server as long as the command runner and the bot has the correct permission.
Purge deletes the specified amount of messages in the channel that the command is ran in as long as it is not over 100 messages and the messages aren't over 14 days old.