Update app.py, bot.py, templates/edit_welcome.html, templates/select_guild.html Create static/style.css
Updated app.py to fix some issues with saving the welcome message to the database. Also enabled debugging when running the app. Update bot.py to fix some issues saving to the database and to add the vent which triggers when a new user joins the discord server.
Created static/style.css which contains some styling for the dashboard this is not final.
updated all templates just adding the css files to the header tag.