Update app.py, bot.py, requirements.txt. Created cogs/fun.py
Updated app.py and bot.py to change the token from the production token to the development token. Same with the databases. Updated requirements.txt to add the discord-games library for the game commands. Created cogs/fun.py which contains 2 of the games from the library. Pushing to main bot soon.