Update bot.py, cogs/moderation.py cogs/server.py. Create utils/generate_punishment_id.py
Updated bot.py, to stop sending the verification message on start up (just for test bot). Added a new punishment ID system in cogs/moderation.py. Changed the emoji when the verify and welcome is set up in cogs/server.py. And created utils/generate_punishment_id.py so it generates the punishment id in a function.