Update cogs/bot.py, cogs/moderation.py
Added a new command in cogs/bot.py that shows the number of guilds the bot is in and the number of users that the bot oversees. Added a new command in cogs/moderation.py called nickname which allows you to change the nickname of the specified user as long as the command runner has manage_nicknames permission.