Update moderation.py, bot.py
Changing how the data is stored in the data base. It all goes to one collection called punishments and there is a type stored. Punishment lookup command added for people to look up punishments through the ID.
Updated the bot.py collections to follow the same as the moderation. Pushing to production now.